1109 Progress

+++ src/main/front-end/app_board/src/Article/FileForm.js
@@ -15,26 +15,29 @@ function Files({files, user, createdId}) {
         }
     }
 
-    return(
-        <div style={{textAlign: "left"}}>
-            {Array.from(files).map((file, index) => {
-                return (<div key={index}>
-                    <span>{file.origin_name}  &nbsp;&nbsp; {Number(file.file_size).toFixed(2)} MB </span>
-                    {canChange(user, createdId) && 
-                        <input type={"image"} src={require("../Icon/remove.png").default} alt={"icon"}
-                                style={{width:"20px", height:"20px", objectFit: "fill", verticalAlign: "middle", marginLeft: "10px"}}
-                                onClick={() => {deleteFile(file.id, file.origin_name)}} />
-                    }
-                    {!_.isEmpty(user) && 
-                        <input type={"image"} src={require("../Icon/download.png").default} alt={"icon"}
-                        style={{width:"20px", height:"20px", objectFit: "fill", verticalAlign: "middle", marginLeft: "10px"}}
-                        onClick={() => {Download(resource, "download", file.id, file.origin_name)}} />
-                    }
-                     <br/> 
-                </div>)
-            })}
-        </div>
-    )
+    if(_.isEmpty(files)){ <div> Loading ...</div>}
+    else{
+        return(
+            <div style={{textAlign: "left"}}>
+                {Array.from(files).map((file, index) => {
+                    return (<div key={index}>
+                        <span>{file.origin_name}  &nbsp;&nbsp; {Number(file.file_size).toFixed(2)} MB </span>
+                        {canChange(user, createdId) && 
+                            <input type={"image"} src={require("../Icon/remove.png").default} alt={"icon"}
+                                    style={{width:"20px", height:"20px", objectFit: "fill", verticalAlign: "middle", marginLeft: "10px"}}
+                                    onClick={() => {deleteFile(file.id, file.origin_name)}} />
+                        }
+                        {!_.isEmpty(user) && 
+                            <input type={"image"} src={require("../Icon/download.png").default} alt={"icon"}
+                            style={{width:"20px", height:"20px", objectFit: "fill", verticalAlign: "middle", marginLeft: "10px"}}
+                            onClick={() => {Download(resource, "download", file.id, file.origin_name)}} />
+                        }
+                         <br/> 
+                    </div>)
+                })}
+            </div>
+        )
+    }
 }
 
 export default Files;
